FAzendo a busca por raio funcionar

+++ src/main/java/br/com/unigranrio/matafome/infra/repositorios/NegocioRepositorioImpl.java
@@ -35,7 +35,7 @@ public class NegocioRepositorioImpl extends RepositorioAbstrato implements Negoc
 	public List<Negocio> obterTodosDentroDoRaio(double raio, LatLng coordenadas) {
 		List<Negocio> barracas = new ArrayList<Negocio>();
 		
-		String query = "SELECT * FROM barraca WHERE " + (raio/1000) + " >= public.geodistance(" + coordenadas.lat + ", " + coordenadas.lng + ", latitude, longitude)";
+		String query = "SELECT * FROM barraca WHERE " + (double)(raio/1000.00) + " >= public.geodistance(" + coordenadas.lat + ", " + coordenadas.lng + ", latitude, longitude)";
 		
 		try {
 			openConnection();
